Allow instances to share a database, data is now scoped to a folder instead of an instance, fix async folder detection, graph data support for all methods, code quality improvements, export Gun for brevity.
Level would sometimes throw an error because Gun instances were trying to open a database that was already open. This has been fixed, and now Gun instances can share and write to the same folder (as they link up to the same level controller).

In the last version, I had this silly idea to scope data to an instance inside of the target folder. Yeah. We changed that.

The last version had another bug where I was trying to detect whether a folder existed asynchronously and, if not, create it asynchronously. That caused some major issues. After reading more documentation, I was able to remove all folder logic completely, leaving the asynchronous bits to level. This will probably be a future feature where I'll let the folder option contain variable depth and blaze a path to it.

In the previous version, `.key` and `.get` weren't holding graph-style data. That's been fixed, so now your key can refer to a collection, and `.get` won't send an error telling of bad data.

The code before would've made a maggot puke. Now they should find it tolerable.

And now, Gun is exported for brevity. You can require `gun-level` and save it as the Gun constructor, instead of making two `require` calls. Yay!