Tests failing: Combine "key" with "put", rewrite put/get for gun@0.3 upgrades, gun.set backfill, upgrade hook-patcher to wire compatibility, request batching in .put, numerous minor code improvements.
In gun version 0.3, `.put` and `.key` are merged into one system. There are required modifications to make that system work, and so far tests are still failing. So far .put works in theory, but intermittently in the tests.

.get and .put have been significantly simplifed by removing the key logic.

Since gun is no longer packaged with chain.set, it is now included until all tests are passing. After that, it will be removed and the tests modified to comply.

patch.js in utils has been modified to comply to the wire spec. Instead of returning an object with a "hooks" property, it's named "wire" instead, now only carrying get and put.

In .put, each node in a graph is pushed to an array, which is then sent as a batch process in level. This allows level to more intelligently manage the data, while also reducing the complexity of counting persisted objects and avoiding possible race conditions in synchronous levelUP implementations.

The code is now a teensy bit prettier.