Radical new rewrite
In response to issue #6, leveldown is no longer pre-packaged with
gun-level (or levelup for that matter). Instead, it's the users
responsibility to set up a level instance and pass it to the gun
constructor. Here are a few reasons why:
 - It gives the developer more freedom.
 - It's easier to understand. Instead of obscuring everything behind a
   wall of configuration, it exposes the interfaces directly to the
developer.
 - There are no implicit level deps (like levelDOWN, which fails the
   install on windows machines).
 - It makes setup simpler, since plugin examples assume you have direct
   access to levelup.
This commit removes nearly all configuration options (like blaze or
share), leaving that to the developer, since many options aren't even
remotely applicable in a browser enviromnent, and it makes little sense
to bundle them into a small adapter library like gun-level.